+++ src/sapper_preloading_b.d.ts
@@ -1,11 +1,11 @@
 import { B } from '@ctx-core/object';
-import { Writable } from '@ctx-core/store';
+import { Writable$ } from '@ctx-core/store';
 export declare const sapper_preloading_b: sapper_preloading_b_T;
 export declare type $sapper_preloading_T = object;
 export declare type $sapper_preloading_type = $sapper_preloading_T;
-export declare type $maybe_sapper_preloading_T = null | $sapper_preloading_T;
+export declare type $maybe_sapper_preloading_T = undefined | $sapper_preloading_T;
 export declare type $maybe_sapper_preloading_type = $maybe_sapper_preloading_T;
-export interface sapper_preloading_T extends Writable<$maybe_sapper_preloading_T> {
+export interface sapper_preloading_T extends Writable$<$maybe_sapper_preloading_T> {
 }
 export declare type sapper_preloading_type = sapper_preloading_T;
 export declare type sapper_preloading_b_T = B<sapper_preloading_T>;