+++ src/sapper_session_b.d.ts
@@ -1,9 +1,9 @@
 import { B } from '@ctx-core/object';
-import { Writable } from '@ctx-core/store';
+import { Writable$ } from '@ctx-core/store';
 export declare const sapper_session_b: sapper_session_b_T;
-export declare type $sapper_session_T = null | object;
+export declare type $sapper_session_T = undefined | object;
 export declare type $sapper_session_type = $sapper_session_T;
-export interface sapper_session_T extends Writable<$sapper_session_T> {
+export interface sapper_session_T extends Writable$<$sapper_session_T> {
 }
 export declare type sapper_session_type = sapper_session_T;
 export interface sapper_session_b_T extends B<sapper_session_T> {